+++ packages/fetch-router/src/lib/form-action.ts
@@ -2,6 +2,22 @@ import type { RoutePattern } from '@remix-run/route-pattern'
 
 import { createRoutes } from './route-map.ts'
 import type { BuildRouteMap } from './route-map.ts'
+import type { RequestMethod } from './request-handler.ts'
+
+export interface FormActionOptions {
+  /**
+   * The method the `<form>` uses to submit to the action route.
+   * Default is `POST`.
+   */
+  submitMethod?: RequestMethod
+  /**
+   * Custom names to use for the `index` and `action` routes.
+   */
+  routeNames?: {
+    index?: string
+    action?: string
+  }
+}
 
 /**
  * Create a route map with `index` (`GET`) and `action` (`POST`) routes, suitable
@@ -10,19 +26,44 @@ import type { BuildRouteMap } from './route-map.ts'
  *
  * @param pattern The route pattern to use for the form and its submit action
  */
-export function createFormAction<P extends string>(
+export function createFormAction<P extends string, const O extends FormActionOptions>(
   pattern: P | RoutePattern<P>,
-): BuildFormActionMap<P> {
+  options?: O,
+): BuildFormActionMap<P, O> {
+  let submitMethod = options?.submitMethod ?? 'POST'
+  let indexName = options?.routeNames?.index ?? 'index'
+  let actionName = options?.routeNames?.action ?? 'action'
+
   return createRoutes(pattern, {
-    index: { method: 'GET', pattern: '/' },
-    action: { method: 'POST', pattern: '/' },
-  })
+    [indexName]: { method: 'GET', pattern: '/' },
+    [actionName]: { method: submitMethod, pattern: '/' },
+  }) as BuildFormActionMap<P, O>
 }
 
-type BuildFormActionMap<P extends string> = BuildRouteMap<
+// prettier-ignore
+type BuildFormActionMap<P extends string, O extends FormActionOptions> = BuildRouteMap<
   P,
   {
-    index: { method: 'GET'; pattern: '/' }
-    action: { method: 'POST'; pattern: '/' }
+    [
+      K in O extends { routeNames: { index: infer I } }
+        ? I extends string
+          ? I
+          : 'index'
+        : 'index'
+    ]: {
+      method: 'GET'
+      pattern: '/'
+    }
+  } & {
+    [
+      K in O extends { routeNames: { action: infer A } }
+        ? A extends string
+          ? A
+          : 'action'
+        : 'action'
+    ]: {
+      method: O extends { submitMethod: infer M } ? M : 'POST'
+      pattern: '/'
+    }
   }
 >